Fix failing cross-policy-versioning functional test
Since [1] the functional tests for cross-policy
object versioning fails with an error. These tests are
skipped in all CI test jobs because none of them have
more than one policy enabled. However, when running
tests against a system with multiple policies the tests
will not skip and the error is raised.

[1] Related-Change: Ifebc1c3ce558b1df9e576a58a4100f2219dfc7e7

Related-Bug: #1629234

Change-Id: I48028928bc996252a31cbf3abfd2b5e67a74dc95